requirements.txt
fastmcp==2.10.4
google-cloud-retail==2.3.0
python-dotenv==1.1.1
pytest
pytest-mock